refactor: (#62) 게임 도메인 타입 정의 shared 레이어로 이동
