--- src/entities/game/model/types.ts
@@ -1,4 +1,0 @@
-export * from './drawing';
-export * from './phases';
-export * from './player';
-export * from './room';
