+++ src/features/game/ui/InventoryPanel.tsx
@@ -1,8 +1,8 @@
 import { ChevronDownIcon, ChevronUpIcon } from '@heroicons/react/24/outline';
 
-import type { GameItem, InventoryUIItem } from '@/entities/game';
 import { ItemIcon } from '@/entities/game';
 import { cn } from '@/shared/lib';
+import type { GameItem, InventoryUIItem } from '@/shared/model';
 import { COLORS } from '../constants/game';
 import { useInventory } from './useInventory';
