+++ src/shared/model/drawing.ts
