+++ src/shared/model/index.ts
@@ -1,1 +1,5 @@
+export * from './drawing';
+export * from './phases';
+export * from './player';
+export * from './room';
 export * from './types';
