+++ src/shared/model/player.ts
