+++ src/shared/model/room.ts
