+++ src/widgets/game-drawing/model/useGameState.ts
@@ -1,8 +1,8 @@
 import { useMemo } from 'react';
 
-import type { DrawingContext, Player } from '@/entities/game';
 import { useAuthStore } from '@/entities/user';
 import { MOCK_GAME_SELECTING } from '@/mocks/game.mock';
+import type { DrawingContext, Player } from '@/shared/model';
 
 export const useGameState = () => {
   const user = useAuthStore((state) => state.user);
